+++ tests/testthat/setup.R
@@ -1,0 +1,1 @@
+write_lines("I should be automatically deleted", "DELETE-ME")