+++ tests/testthat/teardown.R
@@ -1,0 +1,1 @@
+unlink("DELETE-ME")